Allow customizing backtrace printing beyond the feature flag
This adds a builder 'Config' for customizing the panic hook, which
currently only has one option, 'backtrace_mode'. There are three
modes:

- Off: no backtrace
- Unresolved: addresses only
- Resolved: addresses and symbols if possible (default)

The builder is implemented such that if symbol resolution is not
requested, that part of the 'backtrace' crate can be eliminated as
dead code if the client is compiled with LTO (thin or full).